+++ Plugin~/WebRTCPlugin/Codec/NvCodec/NvEncoder.cpp
@@ -309,7 +309,7 @@ namespace webrtc
     }
 
     //entry for encoding a frame
-    bool NvEncoder::EncodeFrame(int64_t timestamp_us, uint32_t* priorityArray)
+    bool NvEncoder::EncodeFrame(int64_t timestamp_us, float* priorityArray)
     {
         UpdateSettings();
         uint32 bufferIndexToWrite = frameCount % bufferedFrameNum;
